requirements.txt
fastapi[standard]
requests
sqlmodel
python-dotenv
APScheduler
pyserial
aiosqlite